create a file for useReactSelectStyles hook
