+++ packages/webapp/src/components/Form/Unit/index.jsx
@@ -13,7 +13,7 @@
  *  GNU General Public License for more details, see <<https://www.gnu.org/licenses/>.>
  */
 
-import React, { useMemo } from 'react';
+import React from 'react';
 import styles from './unit.module.scss';
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
@@ -22,12 +22,12 @@ import { Cross } from '../../Icons';
 import { useTranslation } from 'react-i18next';
 import { integerOnKeyDown, numberOnKeyDown, preventNumberScrolling } from '../Input';
 import Select from 'react-select';
-import { styles as reactSelectDefaultStyles } from '../ReactSelect';
 import { area_total_area } from '../../../util/convert-units/unit';
 import Infoi from '../../Tooltip/Infoi';
 import { Controller } from 'react-hook-form';
 import { ReactComponent as Leaf } from '../../../assets/images/signUp/leaf.svg';
 import useUnit from './useUnit';
+import useReactSelectStyles from './useReactSelectStyles';
 
 const getOnKeyDown = (measure) => {
   switch (measure) {
@@ -38,64 +38,8 @@ const getOnKeyDown = (measure) => {
   }
 };
 
-const DEFAULT_REACT_SELECT_WIDTH = 80;
 const DEFAULT_SELECT_ARROW_ICON_WIDTH = 20;
 
-const getReactSelectWidth = (measure) => {
-  if (measure === 'time') return 93;
-  return DEFAULT_REACT_SELECT_WIDTH;
-};
-
-const useReactSelectStyles = (disabled, { reactSelectWidth = DEFAULT_REACT_SELECT_WIDTH } = {}) => {
-  return useMemo(
-    () => ({
-      ...reactSelectDefaultStyles,
-      container: (provided, state) => ({
-        ...provided,
-      }),
-      control: (provided, state) => ({
-        display: 'flex',
-        border: `none`,
-        boxShadow: 'none',
-        boxSizing: 'border-box',
-        borderRadius: '4px',
-        height: '48px',
-        paddingLeft: '0',
-        fontSize: '16px',
-        lineHeight: '24px',
-        color: 'var(--fontColor)',
-        background: 'transparent',
-      }),
-      valueContainer: (provided, state) => ({
-        ...provided,
-        padding: '0',
-        width: `${reactSelectWidth - 19}px`,
-        display: 'flex',
-        justifyContent: 'center',
-        background: disabled ? 'var(--inputDisabled)' : 'inherit',
-      }),
-      singleValue: (provided, state) => ({
-        fontSize: '16px',
-        lineHeight: '24px',
-        color: state.isDisabled ? 'var(--grey600)' : 'var(--fontColor)',
-        fontStyle: 'normal',
-        fontWeight: 'normal',
-        fontFamily: '"Open Sans", "SansSerif", serif',
-        overflowX: 'hidden',
-      }),
-      placeholder: () => ({
-        display: 'none',
-      }),
-      dropdownIndicator: (provided, state) => ({
-        ...provided,
-        display: state.isDisabled ? 'none' : 'flex',
-        padding: ' 14px 0 12px 0',
-        transform: 'translateX(-4px)',
-      }),
-    }),
-    [disabled, reactSelectWidth],
-  );
-};
 const Unit = ({
   disabled = false,
   classes = { container: {} },
@@ -159,8 +103,7 @@ const Unit = ({
     max,
   });
 
-  const reactSelectWidth = getReactSelectWidth(measure);
-  const reactSelectStyles = useReactSelectStyles(disabled, { reactSelectWidth });
+  const { reactSelectStyles, reactSelectWidth } = useReactSelectStyles(disabled, { measure });
 
   return (
     <div className={clsx(styles.container)} style={{ ...style, ...classes.container }}>
@@ -212,7 +155,7 @@ const Unit = ({
         <Controller
           control={control}
           name={displayUnitName}
-          render={({ field: { onChange, onBlur, value, name, ref } }) => (
+          render={({ field: { onChange, onBlur, value, ref } }) => (
             <Select
               data-cy="unit-select"
               onBlur={onBlur}
